+++ app/domain/errors.rb
@@ -11,12 +11,12 @@ module Errors
   module Group
     DuplicateMember = ::Util::TrackableErrorClass.new(
       msg: "The '{0}' resource (kind='{1}') is already a member of the '{2}' group",
-      code: "CONJ00180E"
+      code: "CONJ00180W"
     )
 
     ResourceNotMember = ::Util::TrackableErrorClass.new(
       msg: "The '{0}' resource (kind='{1}') is not a member of the '{2}' group",
-      code: "CONJ00181E"
+      code: "CONJ00181W"
     )
   end
 
@@ -43,7 +43,7 @@ module Errors
 
     APIHeaderMissing = ::Util::TrackableErrorClass.new(
       msg: "The api belongs to v2 APIs but it missing the version \"application/x.secretsmgr.v2+json\" in the Accept header",
-      code: "CONJ00194E"
+      code: "CONJ00194W"
     )
 
     RequiredResourceMissing = ::Util::TrackableErrorClass.new(